+++ app/src/Navigation/MainNavigation/MyHelpRequestsNavigator/index.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
-import { Image } from 'react-native';
 import OnGoingHelps from '../../../pages/helpPages/MyRequests/onGoing';
 import DoneHelps from '../../../pages/helpPages/MyRequests/doneHelps';
 import colors from '../../../../assets/styles/colorVariables';
 import fonts from '../../../../assets/styles/fontVariable';
 import HelpDescription from '../../../pages/helpPages/helpDescription';
-import backImage from '../../../../assets/images/back.png';
+import headerStyle from '../MainStackHeaderStyle';
 
 const MyRequestsTab = createMaterialTopTabNavigator();
 const stack = createStackNavigator();
@@ -47,33 +46,6 @@ const tabTopBarOptions = {
         backgroundColor: colors.light,
         padding: 2,
     },
-};
-
-const headerStyle = {
-    headerBackImage: () => (
-        <Image
-            source={backImage}
-            style={{
-                flex: 1,
-                resizeMode: 'contain',
-                width: 10,
-                marginLeft: 5,
-            }}
-        />
-    ),
-    headerStyle: {
-        height: 90,
-        backgroundColor: colors.primary,
-        elevation: 0,
-    },
-    headerTitleStyle: {
-        ...fonts.title,
-        color: colors.light,
-        fontFamily: 'montserrat-medium',
-        marginHorizontal: 30,
-    },
-    headerTintColor: colors.light,
-    headerTitleAlign: 'center',
 };
 
 export default MyRequestsNavigation;
