+++ app/src/Navigation/MainNavigation/MyOfferedHelpNavigator/index.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import on_goingGivenHelp from '../../../pages/helpPages/givenHelps/on_going';
 import finishedGivenHelp from '../../../pages/helpPages/givenHelps/finished';
-import { Image } from 'react-native';
-import backImage from '../../../../assets/images/back.png';
 import HelpDescription from '../../../pages/helpPages/helpDescription';
 import colors from '../../../../assets/styles/colorVariables';
 import fonts from '../../../../assets/styles/fontVariable';
+import headerStyle from '../MainStackHeaderStyle';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 import { createStackNavigator } from '@react-navigation/stack';
 
@@ -44,33 +43,6 @@ const HelpTopBar = () => (
 );
 
 export default HelpTopBar;
-
-const headerStyle = {
-    headerBackImage: () => (
-        <Image
-            source={backImage}
-            style={{
-                flex: 1,
-                resizeMode: 'contain',
-                width: 10,
-                marginLeft: 5,
-            }}
-        />
-    ),
-    headerStyle: {
-        height: 90,
-        backgroundColor: colors.primary,
-        elevation: 0,
-    },
-    headerTitleStyle: {
-        ...fonts.title,
-        color: colors.light,
-        fontFamily: 'montserrat-medium',
-        marginHorizontal: 30,
-    },
-    headerTintColor: colors.light,
-    headerTitleAlign: 'center',
-};
 
 const tabTopBarOptions = {
     style: {
